:recycle: Extract PartsCollector into shared module
Extract common PartsCollector struct and parts_to_ruby_array helper
from datetime_format.rs and number_format.rs into a dedicated
parts_collector.rs module.
